chore: update tarantool-module as well as sbroad
sbroad needs to be updated to the version that supports newer
tarantool-module

sbroad support for new module version was merged here:
https://git.picodata.io/picodata/picodata/sbroad/-/merge_requests/355

+++ src/util.rs
@@ -566,7 +566,7 @@ pub type CheckIsSameType<L, R> = <L as IsSameType<L, R>>::Void;
 /// A helper struct to enforce that a function must not yield. Will cause a
 /// panic if fiber yields are detected when drop is called for it.
 pub struct NoYieldsGuard {
-    csw: i32,
+    csw: u64,
 }
 
 #[allow(clippy::new_without_default)]
